feat(api): add featured workflows query to craft module router
refactor(desktop): replace lazy loading of layout index route with direct import

refactor(web): use api.craft.module.featured for fetching featured workflows in explore page

chore(web): remove redundant getFeaturedWorkflows function from explore actions

refactor(components): remove template-list component to clean up unused code
fix(types): correct import and declaration of gql.tada module in wordpress-graphql.d.ts
chore(tsconfig): update include paths in tsconfig.json to add workflow-list.tsx

feat(ui): add WorkflowList and WorkflowItem components

Introduce WorkflowList and WorkflowItem components to display a list
of workflows with their details, including project information and
usage statistics.
