+++ apps/web/src/app/(dashboard)/explore/actions.ts
@@ -19,21 +19,3 @@ export const getUserProjects = async () => {
     },
   });
 };
-
-export const getFeaturedWorkflows = async () => {
-  return await db.query.workflow.findMany({
-    where: (workflow, { eq, and }) => and(eq(workflow.public, true)),
-    orderBy: (workflow, { desc }) => desc(workflow.updatedAt),
-    limit: 20,
-    with: {
-      project: true,
-      versions: {
-        columns: {
-          version: true,
-        },
-        orderBy: (version, { desc }) => desc(version.version),
-        limit: 1,
-      },
-    },
-  });
-};
