+++ apps/web/tsconfig.json
@@ -20,6 +20,10 @@
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json",
     "module": "esnext"
   },
-  "include": [".", ".next/types/**/*.ts"],
+  "include": [
+    ".",
+    ".next/types/**/*.ts",
+    "../../packages/ui/src/views/workflow-list.tsx"
+  ],
   "exclude": ["node_modules"]
 }
